#43 make common decoration
