+++ presentation/src/main/java/com/jslee/presentation/feature/upcoming/UpComingFragment.kt
@@ -3,10 +3,11 @@ package com.jslee.presentation.feature.upcoming
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.jslee.core.ui.base.view.BaseFragment
+import com.jslee.core.ui.decoration.CommonItemDecoration
+import com.jslee.core.ui.decoration.LayoutType
 import com.jslee.core.ui.model.PaddingValues
 import com.jslee.presentation.R
 import com.jslee.presentation.databinding.FragmentUpComingBinding
-import com.jslee.presentation.feature.home.decoration.HomeUpComingItemDecoration
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
@@ -36,7 +37,7 @@ class UpComingFragment : BaseFragment<FragmentUpComingBinding>(R.layout.fragment
     private fun initRecyclerView() = with(binding.rvUpComing) {
         adapter = upComingPagingAdapter
         val paddingValues = PaddingValues(2, 2, 2, 2)
-        addItemDecoration(HomeUpComingItemDecoration(paddingValues))
+        addItemDecoration(CommonItemDecoration(paddingValues, LayoutType.GRID))
     }
 
     override fun observeStates() {
